// src/webparts/siteCatalog/components/SiteCatalog.module.scss.d.ts
declare const styles: {
    siteCatalog: string;
    container: string;
    row: string;
    column: string;
    'ms-Grid': string;
    title: string;
    subTitle: string;
    description: string;
    button: string;
    label: string;
    floatRight: string;
    'ms-SearchBoxContainer': string;
    floatIconRight: string;
    alignCenter: string;
    moreButton: string;
    groupIcons: string;
};
export default styles;